requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10
web3>=7.0.0
eth-account>=0.11.0
httpx>=0.26.0
websockets>=12.0
cryptography>=42.0.0
python-dotenv>=1.0.0
click>=8.0.0
py-clob-client>=2.0.0

